requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0
anthropic>=0.28.0
celery>=5.3.0
redis>=5.0.0
reportlab>=4.1.0
Pillow>=10.3.0
opencv-python-headless>=4.9.0  # optional: for video frame extraction
